Split trait interfaces into modules
